+++ src/Facades/Translatable.php
@@ -10,6 +10,7 @@ use Illuminate\Support\Facades\Facade;
  * @method static void fallback(?string $fallbackLocale = null, ?bool $fallbackAny = false, $missingKeyCallback = null)
  * @method static void allowNullForTranslation(bool $allowNullForTranslation = true)
  * @method static void allowEmptyStringForTranslation(bool $allowEmptyStringForTranslation = true)
+ * @method static void allowEvent(bool $allowEvent = true)
  */
 class Translatable extends Facade
 {